Crate fck

This is the source code for fck(pure). This is the basis for all other fck flavours

Modules

ast

Contains the lexer and parser for fck. There are language agnostic and as such cannot use regex.

bases

Base file with all the “building blocks” of everything else in

config_file

Config file handling

err_wrn

llvm

Everything LLVM related

nodes

AST Nodes

shell

Shell file using JIT compilation

tokens

Token enum

Functions

get_file_name

main

Reads the local config file and parses arguments accordingly

run_file