Struct fck::ast::Lexer

pub struct Lexer {
    split_text: Vec<char>,
    current_pos: Position,
    char_index: usize,
    current_char: char,
    pub(crate) keywords: Keywords<'static>,
    pub(crate) keyword_code: String,
}

Lexer for fck

This works on a (8-bit) character by character basis. 16-bit characters are still in production so be patient

Fields
split_text: Vec<char>

Input text split into a vector of characters

current_pos: Position

Current position of the “pointer”

char_index: usize

The current index of the current character in split_text

current_char: char

the current character the pointer is pointing to

keywords: Keywords<'static>

The current keyword struct words are being checked against

keyword_code: String

The string associated with the current keywords

Implementations

impl Lexer

pub fn new(
full_text: String,
keywords: Keywords<'static>,
keyword_code: String
) -> Lexer

Returns a new lexer with all fields completed appropriately

fn advance(&mut self)

Advances the self.current_pos by 1 and updates the character it points to

pub fn make_tokens(&mut self) -> Result<Vec<Token>, Error>

Primary function of the lexer. Iterates over the input text and attempts to turn it into a vector of tokens. If an error occurs, an Err(Error) is returned

fn make_number(&mut self) -> Token

Called by self.make_tokens when a number is found. Iterates over characters until a second ‘.’ or non-numeric character is found, at which point the function returns the appropriate token

fn make_identifier(&mut self) -> Token

Called by self.make_tokens when a alphabetic character is found. This function either returns a TT_KEYWORD or TT_IDENTIFIER token depending on the string and current keywords

fn make_loop_identifier(&mut self) -> Result<Token, Error>

Called by self.make_tokens when a ‘@’ character is found. This advances and calls self.make_identifier, taking the value of the token returned and placing it into a new TT_AT token

fn make_string(&mut self, starting_character: char) -> Result<Token, Error>

Called by self.make_tokens when an opening string delimiter is found. Parses the string and returns a TT_STRING token containing the string literal in an Ok(Token). Returns an Err(Error) if no closing delimiter is found.

fn make_not_equals(&mut self) -> Result<Token, Error>

Called by self.make_tokens when a ‘!’ is found. Will check for a “!=”, “!!”, or character after it otherwise an Err(Error) is returned.

fn make_equals(&mut self) -> Result<Token, Error>

Called by self.make_tokens when a ‘=’ is found. Will check for a “==” otherwise an Err(Error) is returned

fn single_double_token(
&mut self,
second_char: char,
single_type: TokType,
double_type: TokType
) -> Result<Token, Error>

General function called by self.make_tokens when we want to check for one of two tokens that start with the same character with the second token being followed by another character such as // or <=

fn make_set(&mut self) -> Result<Token, Error>

Called by self.make_tokens when a ‘:’ is found. This parses one or two more characters to decide what kind of assignment operation is being performed. It makes use of the ordering of token type values to make the code simpler

fn skip_comment(&mut self)

Called by self.make_tokens when a comment is found. This skips the comment. Big surprise