Module fck::llvm

Everything LLVM related

This generates LLVM IR code. This file does not contain code to handle compilation to executables, or JIT compilation. They’re handled in compiling.rs and interpreter.rs

Constants

LLVM_FALSE

False constant of type LLVMBool

LLVM_TRUE

True constant of type LLVMBool

Functions

build_arith_op⚠

Builds a ASTNodeType::ArithOp AST node to the module

build_ast

Builds the ASTs to the module

build_comp_op⚠

build_if⚠

build_var_assign⚠

Builds a ASTNodeType::VarAssign AST node to the module

get_default_target_triple

Gets the default target string

init⚠

Initialise LLVM

ir_to_module

Main function of the file. Turns the asts into an LLVM module

llvm_int⚠

Useful function to make an integer

object_to_executable

convert the object file into an executable using clang

to_object_file

Builds a module to an object file